Make HMR work on windows
On windows there are inconsistencies regarding the file paths because of the different path seperator. This leads to the dependencies variable to be something like this:
```
Compilable Files Map(1) {
  '.../src/Main.elm' => Set(1) {     
    '...\\src\\Translations.elm'
  }
}
```

The Set.has call does not take this inconsistency into account and always fails.
The smallest fix I could find is the application of path.normalize. I'm pretty sure this should not break other operating systems and it fixes the problem on windows.